+++ src/app/layout.tsx
@@ -7,7 +7,7 @@ import { Header } from '@/components/Header';
 export const metadata: Metadata = {
   title: 'Archie Biddiscombe',
   description:
-    'Archie is an Associate Software Engineer with a specialism in geospatial technologies.',
+    'Archie is an Associate Software Engineer (React frontend development) based in the UK and currently working for Ordnance Survey',
 };
 
 interface RootLayoutProps {
@@ -20,13 +20,11 @@ export default function RootLayout(p: RootLayoutProps) {
       <head>
         <meta property='og:type' content='website' />
         <meta property='og:url' content='https://abiddiscombe.dev/' />
-        <meta property='og:title' content="Archie Biddiscombe's Site" />
-        <meta property='og:description' content={metadata.description || ''} />
+        <meta property='og:title' content="Archie Biddiscombe's Website" />
         <link rel='icon' href='./favicon/favicon.ico' />
         <link rel='icon' type='image/png' sizes='32x32' href='./favicon/favicon-32x32.png' />
         <link rel='icon' type='image/png' sizes='16x16' href='./favicon/favicon-16x16.png' />
       </head>
-
       <body className='p-6 pt-10 h-screen sm:grid sm:place-items-center'>
         <main className='max-w-xl'>
           <Header />