+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 4.0.0 - February 2024
 
 1. feat: Site redesign. \*
+2. chore: Update site description to match content.
 
 \* - This commit includes various changes to site structure and copy.